Revert the removal of parse support for edict/edict2
The edict/edict2 parser was used by submit.py to parse edict2 formatted
entries submitted with the "&j=" url parameter.  That ability in turn was
used by wwwjdic for aiding bulk submissions and for promoting related words
found in alternative dictionaries into the main jmdict dictionary.  This
commit restores that ability by reverting the commits that removed it.

This is a squashed commit of the git reverts of commits:
  221028-4544535 (from xng): Removed parse support for edict/edict2 (cgi)
  220828-c2b0818 (from xng): Removed parse support for edict/edict2 (data)
  220828-a3ce097 (ftom xng): Removed parse support for edict/edict2