Issue #254 Stop forks... Add various checks and warning messages to avoid unintentional forks.
Note: this commit adds a column 'leaf' to the public.esum view in jmtest01.sql, but does not include a migration to repair that view on upgrade. TODO.

This change limits the ability to create forks, provides mechanisms for easily reviewing changes in the event of concurrent modification, and adds some additional error checking in the presense of fork conflicts.

* Add an admin-only checkbox to the "Edit" screen to enable the old forking behavior (allow submissions to fork). Otherwise...

* Submission can only occur against a leaf entry

* A submission against a non-leaf entry will fast-forward automatically to the leaf(parent eid is replaced at the beginning of the submission process).  A warning message will be displayed, and "skipped" data (which will be clobbered) will be highlighted in red in the edit history on the submission-confirmation page..

* Approvals and Rejections against a non-leaf entry will fast-forward (as above) if there are no forks (They will still clobber).  If there are forks on the sequence, an error message will be displayed with links to the head of the current branch (if unambiguous), along with the remaining leafs.

* Additional checks are added before "submit".  If old-style-forking not enabled, and the entry that would be appended to is no longer a leaf (submitting would create a fork), an error message appears instead. (This occurs when someone else submits an edit between [Next] and [Submit]).  The user can use the 'back' button to re-submit the initial edit, which will fast-forward if possible.  This final check occurs in a transaction, and is race-condition safe.

This commit also adds display change to the search-results page. Unapproved-entry items, which were previously marked in red with an asterisk (*), are now marked with *either* an asterisk, or a hyphen (-).  Asterisk indicates "leaf" entry, while hyphen indicates an intermediate edit which should not usually be operated on directly. These intermediate edits do, however, provide a clean, complete copy of a user-submission, so they could be handy in recovering accidentally-clobbered data.  They can also be forked off of, if desired.

No effort is made to "merge" or otherwise reason about the intentions involved in clobbering data.  This just keeps everything on one thread/edit chain, where manual fixes can still be made if needed. As long as a 'rejection' or 'approval' has not occurred, the full version of every submission remains accessible in the search listing (with non-leaf entries marked with a hyphen(-) in place of an asterisk.